+++ core/types/exe.ts
@@ -89,6 +89,7 @@ export type ExeSubtype =
   | 'exeWhen'
   | 'exeForeach'
   | 'exeFor'
+  | 'exeLoop'
   | 'exeResolver'
   | 'exeBlock';